Remove heap buffer overflow and double free when removing a word
This commit fixes a heap buffer overflow and a double free exploit caused
by removing a word. We now check if the index of the word to be removed
is negative and we only move valid addresses of words within the words
array. We also only free used words in the words array.